apps/azure-finops-exporter/requirements.txt
azure-identity==1.19.0
prometheus-client==0.21.1
requests==2.32.3
